+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Added load balancer server pool [#205](https://github.com/vmware/go-vcloud-director/pull/205)
 * Added load balancer application profile [#208](https://github.com/vmware/go-vcloud-director/pull/208)
 * Added load balancer application rule [#212](https://github.com/vmware/go-vcloud-director/pull/212)
+* Added load balancer virtual server [#215](https://github.com/vmware/go-vcloud-director/pull/215)
 * Added functions for refreshing, getting and update Org VDC [#206](https://github.com/vmware/go-vcloud-director/pull/206)
 * Added VDC meta data create/get/delete functions [#203](https://github.com/vmware/go-vcloud-director/pull/203)
 * Added org user create/delete/update functions [#18](https://github.com/vmware/go-vcloud-director/issues/18)